requirements.txt
django
djangorestframework
django-rest-auth[with_social]
django-model-utils
requests
flake8
drf-url-filters
six
django-cors-headers
mysqlclient
pyspark
whitenoise
gunicorn
